--- src/main/java/lottogame/InputView.java
@@ -1,17 +1,0 @@
-package lottogame;
-
-import java.util.Scanner;
-
-public class InputView {
-	private InputView(){
-	}
-
-	private static Scanner scanner = getScanner();
-
-	public static String readLine() {
-		return scanner.nextLine();
-	}
-	private static Scanner getScanner() {
-		return new Scanner(System.in);
-	}
-}
