--- src/main/java/lottogame/exception/DuplicateLottoNumberException.java
@@ -1,8 +1,0 @@
-package lottogame.exception;
-
-public class DuplicateLottoNumberException extends InvalidLottoFormatException {
-
-	public DuplicateLottoNumberException(String message) {
-		super(message);
-	}
-}
