--- src/main/java/lottogame/exception/InvalidLottoFormatException.java
@@ -1,8 +1,0 @@
-package lottogame.exception;
-
-public class InvalidLottoFormatException extends RuntimeException {
-
-	public InvalidLottoFormatException(String message) {
-		super(message);
-	}
-}
