--- src/main/java/lottogame/exception/UnexpectedMatchException.java
@@ -1,8 +1,0 @@
-package lottogame.exception;
-
-public class UnexpectedMatchException extends RuntimeException{
-
-	public UnexpectedMatchException(String message) {
-		super(message);
-	}
-}
