+++ src/main/java/lottoservice/exception/DuplicateLottoNumberException.java
@@ -1,0 +1,8 @@
+package lottoservice.exception;
+
+public class DuplicateLottoNumberException extends InvalidLottoFormatException {
+
+	public DuplicateLottoNumberException(String message) {
+		super(message);
+	}
+}
