+++ src/main/java/lottoservice/exception/InvalidLottoFormatException.java
@@ -1,0 +1,8 @@
+package lottoservice.exception;
+
+public class InvalidLottoFormatException extends RuntimeException {
+
+	public InvalidLottoFormatException(String message) {
+		super(message);
+	}
+}
