+++ src/main/java/lottoservice/exception/UnexpectedMatchException.java
@@ -1,0 +1,8 @@
+package lottoservice.exception;
+
+public class UnexpectedMatchException extends RuntimeException{
+
+	public UnexpectedMatchException(String message) {
+		super(message);
+	}
+}
