+++ src/main/java/lottoservice/ui/InputView.java
@@ -1,0 +1,17 @@
+package lottoservice.ui;
+
+import java.util.Scanner;
+
+public class InputView {
+	private InputView(){
+	}
+
+	private static Scanner scanner = getScanner();
+
+	public static String readLine() {
+		return scanner.nextLine();
+	}
+	private static Scanner getScanner() {
+		return new Scanner(System.in);
+	}
+}
